+++ crates/zizmor/src/models/coordinate.rs
@@ -59,7 +59,7 @@ impl ActionCoordinate {
 
         // If our coordinate's `uses:` template doesn't match the step's `uses:`,
         // then no usage semantics are possible.
-        if !uses_pattern.matches(uses) {
+        if !uses_pattern.matches(&uses.into()) {
             return None;
         }
 
@@ -401,7 +401,7 @@ pub(crate) enum Usage {
 
 #[cfg(test)]
 mod tests {
-    use std::str::FromStr;
+    use std::str::FromStr as _;
 
     use github_actions_models::common::EnvValue;
     use indexmap::IndexMap;